refactor(constants): Move tool configuration and capabilities functions to a new utility file
